sunxi: nand: Add information to sunxi that it was run from NAND in SPL
As SPL does not know which source to choose when booting U-Boot, choose
NAND if it is capable of doing so.

Signed-off-by: Peter Gielda <pgielda@antmicro.com>
Signed-off-by: Tomasz Gorochowik <tgorochowik@antmicro.com>
Signed-off-by: Mateusz Holenko <mholenko@antmicro.com>
Signed-off-by: Piotr Zierhoffer <pzierhoffer@antmicro.com>

Series-to: u-boot
Series-cc: sunxi, marex
Series-cc: Boris Brezillon <boris.brezillon@free-electrons.com>

Cover-letter:
sunxi: nand: Basic NAND driver with SPL support

This is a basic driver for the sunxi NAND controller for Allwinner A20.
It supports both SPL and U-Boot.

The driver uses DMA for data transfers. It does not support writing.

To enable user reading from NADN, the a20_nandread command was added.
END